Patatap App
This Patatap App is made using paper.js. Every time the user hits a letter key a circle of a specific colour to the key pressed appears, then as the circle fades out it gets smaller and the hue of the colour changes.  This app is the last app I made during the Udemy Web Developer Bootcamp.