ToDo list App
My ToDo list app is made with jQuery, Font Awesome, html5 and css. When the user hovers over a <li> a delete icon appears and disappears when the pointer leaves the <li>. The user can also cross out an <li> by clicking on it or delete by clicking on the delete icon. New Todo's can be clicking the plus sign and hitting enter after writing in the input box.